Move indexing functionality to separate modules
