+++ include/disparity_graph.hpp
@@ -397,176 +397,6 @@ struct DisparityGraph
     );
 };
 
-
-/**
- * \brief Get index of a neighbor for fast access in different data arrays.
- *
- * Pixel is a more complex structure than a number.
- * Tree is a slow structure for our purposes.
- *
- * We can store information associated with pixel and its neighbors
- * in an array,
- * but this needs a calculation of corresponding indices for data access.
- * This function provides these indices.
- *
- * @return
- *  Integer from `[0; ::NEIGHBORS_COUNT - 1]` range
- *  if there should be at least one Edge between two provided pixels.
- *  ::NEIGHBORS_COUNT if the `neighbor` is not a neighbor of the pixel.
- */
-ULONG neighbor_index(
-    struct Pixel pixel,
-    struct Pixel neighbor
-);
-
-/**
- * \brief Check wheter provided Pixel instances
- * in given DisparityGraph are neighbors
- * and exist at all.
- *
- * As it's explained in ::NEIGHBORS_COUNT,
- * there are four possible neighbors.
- *
- * Logic says that
- * top pixels don't have top neighbors,
- * bottom pixels don't have bottom neighbors,
- * right pixels don't have right neighbors
- * and left pixels don't have left pixels.
- *
- * Also, pixels that are not the nearest straight neighbors,
- * are not a neighbors in the DisparityGraph.
- */
-bool neighborhood_exists(
-    const struct DisparityGraph& graph,
-    struct Pixel pixel,
-    struct Pixel neighbor
-);
-
-/**
- * \brief Check existence of provided Pixel instances in given DisparityGraph.
- *
- * See ::neighborhood_exists for more information.
- *
- * This function differs from ::neighborhood_exists
- * just by the second parameter:
- * it takes a neighbor index
- * calculated by ::neighbor_index function.
- */
-bool neighborhood_exists_fast(
-    const struct DisparityGraph& graph,
-    struct Pixel pixel,
-    ULONG neighbor_index
-);
-
-/**
- * \brief Check existence of provided Node.
- *
- * Node exists if position of its pixel doesn't overflow
- * DisparityGraph::right image,
- * corresponding pixel (calculated by the disparity)
- * doesn't overflow the DisparityGraph::left image,
- * and its disparity doesn't overflow
- * the maximal allowed.
- */
-bool node_exists(
-    const struct DisparityGraph& graph,
-    struct Node node
-);
-
-/**
- * \brief Check existence of provided Edge in given DisparityGraph.
- *
- * Similar to ::neighborhood_exists,
- * but also checks constraints imposed on disparities
- * of the neighboring pixels.
- */
-bool edge_exists(
-    const struct DisparityGraph& graph,
-    struct Edge edge
-);
-
-/**
- * \brief Get an index of DisparityGraph::reparametrization element
- * using a Node and an index of its neighbor.
- *
- * The function doesn't check existence of neighbor.
- * You should perform it by yourself
- * using ::neighborhood_exists.
- */
-ULONG reparametrization_index_fast(
-    const struct DisparityGraph& graph,
-    struct Node node,
-    ULONG neighbor_index
-);
-
-/**
- * \brief Get an index of DisparityGraph::reparametrization element
- * using a Node and a Pixel.
- *
- * The function doesn't check existence of neighbor.
- * You should perform it by yourself
- * using ::neighborhood_exists.
- */
-ULONG reparametrization_index(
-    const struct DisparityGraph& graph,
-    struct Node node,
-    struct Pixel neighbor
-);
-
-/**
- * \brief Get an index of DisparityGraph::reparametrization element
- * using an Edge.
- *
- * The function doesn't check existence of neighbor.
- * You should perform it by yourself
- * using ::neighborhood_exists.
- */
-ULONG reparametrization_index_slow(
-    const struct DisparityGraph& graph,
-    struct Edge edge
-);
-
-/**
- * \brief Get a value of DisparityGraph::reparametrization element
- * using a Node and an index of its neighbor.
- *
- * The function doesn't check existence of neighbor.
- * You should perform it by yourself
- * using ::neighborhood_exists.
- */
-FLOAT reparametrization_value_fast(
-    const struct DisparityGraph& graph,
-    struct Node node,
-    ULONG neighbor_index
-);
-
-/**
- * \brief Get a value of DisparityGraph::reparametrization element
- * using a Node and a Pixel.
- *
- * The function doesn't check existence of neighbor.
- * You should perform it by yourself
- * using ::neighborhood_exists.
- */
-FLOAT reparametrization_value(
-    const struct DisparityGraph& graph,
-    struct Node node,
-    struct Pixel neighbor
-);
-
-/**
- * \brief Get a value of DisparityGraph::reparametrization element
- * using an Edge.
- *
- * The function doesn't check existence of neighbor.
- * You should perform it by yourself
- * using ::neighborhood_exists.
- */
-FLOAT reparametrization_value_slow(
-    const struct DisparityGraph& graph,
-    struct Edge edge
-);
-
 /**
  * \brief Calculate penalty of Edge without neighborhood check.
  *
