+++ include/indexing.hpp
@@ -46,4 +46,172 @@ ULONG get_pixel_index(const struct Image& image, struct Pixel pixel);
  */
 ULONG get_pixel_value(const struct Image& image, struct Pixel pixel);
 
+/**
+ * \brief Get an index of DisparityGraph::reparametrization element
+ * using a Node and an index of its neighbor.
+ *
+ * The function doesn't check existence of neighbor.
+ * You should perform it by yourself
+ * using ::neighborhood_exists.
+ */
+ULONG reparametrization_index_fast(
+    const struct DisparityGraph& graph,
+    struct Node node,
+    ULONG neighbor_index
+);
+
+/**
+ * \brief Get an index of DisparityGraph::reparametrization element
+ * using a Node and a Pixel.
+ *
+ * The function doesn't check existence of neighbor.
+ * You should perform it by yourself
+ * using ::neighborhood_exists.
+ */
+ULONG reparametrization_index(
+    const struct DisparityGraph& graph,
+    struct Node node,
+    struct Pixel neighbor
+);
+
+/**
+ * \brief Get an index of DisparityGraph::reparametrization element
+ * using an Edge.
+ *
+ * The function doesn't check existence of neighbor.
+ * You should perform it by yourself
+ * using ::neighborhood_exists.
+ */
+ULONG reparametrization_index_slow(
+    const struct DisparityGraph& graph,
+    struct Edge edge
+);
+
+/**
+ * \brief Get a value of DisparityGraph::reparametrization element
+ * using a Node and an index of its neighbor.
+ *
+ * The function doesn't check existence of neighbor.
+ * You should perform it by yourself
+ * using ::neighborhood_exists.
+ */
+FLOAT reparametrization_value_fast(
+    const struct DisparityGraph& graph,
+    struct Node node,
+    ULONG neighbor_index
+);
+
+/**
+ * \brief Get a value of DisparityGraph::reparametrization element
+ * using a Node and a Pixel.
+ *
+ * The function doesn't check existence of neighbor.
+ * You should perform it by yourself
+ * using ::neighborhood_exists.
+ */
+FLOAT reparametrization_value(
+    const struct DisparityGraph& graph,
+    struct Node node,
+    struct Pixel neighbor
+);
+
+/**
+ * \brief Get a value of DisparityGraph::reparametrization element
+ * using an Edge.
+ *
+ * The function doesn't check existence of neighbor.
+ * You should perform it by yourself
+ * using ::neighborhood_exists.
+ */
+FLOAT reparametrization_value_slow(
+    const struct DisparityGraph& graph,
+    struct Edge edge
+);
+
+/**
+ * \brief Get index of a neighbor for fast access in different data arrays.
+ *
+ * Pixel is a more complex structure than a number.
+ * Tree is a slow structure for our purposes.
+ *
+ * We can store information associated with pixel and its neighbors
+ * in an array,
+ * but this needs a calculation of corresponding indices for data access.
+ * This function provides these indices.
+ *
+ * @return
+ *  Integer from `[0; ::NEIGHBORS_COUNT - 1]` range
+ *  if there should be at least one Edge between two provided pixels.
+ *  ::NEIGHBORS_COUNT if the `neighbor` is not a neighbor of the pixel.
+ */
+ULONG neighbor_index(
+    struct Pixel pixel,
+    struct Pixel neighbor
+);
+
+/**
+ * \brief Get an index of ConstraintGraph::nodes_availability element
+ * using a Node.
+ *
+ * The function doesn't check existence of the Node.
+ * You should perform it by yourself
+ * using ::node_exists.
+ */
+ULONG node_index(const struct DisparityGraph& graph, struct Node node);
+
+/**
+ * \brief Get index of a pixel
+ * to access needed value of LowestPenalties::pixels fast.
+ */
+ULONG pixel_index(
+    const struct DisparityGraph& graph,
+    struct Pixel pixel
+);
+
+/**
+ * \brief Get index of a neighborhood in LowestPenalties::neighborhoods.
+ *
+ * Note that the function doesn't check existence of provided neighborhood.
+ * Use ::neighborhood_exists_fast to make sure that you use it right.
+ */
+ULONG neighborhood_index_fast(
+    const struct DisparityGraph& graph,
+    struct Pixel pixel,
+    ULONG neighbor_index
+);
+
+/**
+ * \brief Get index of a neighborhood given two neighboring pixels.
+ *
+ * Note that the function doesn't check existence of provided neighborhood.
+ * Use ::neighborhood_exists to make sure that you use it right.
+ */
+ULONG neighborhood_index(
+    const struct DisparityGraph& graph,
+    struct Pixel pixel,
+    struct Pixel neighbor
+);
+
+/**
+ * \brief Get index of a neighborhood by arbitrary edge from the neighborhood.
+ *
+ * Note that the function doesn't check existence of provided neighborhood.
+ * Use ::edge_exists to make sure that you use it right.
+ */
+ULONG neighborhood_index_slow(
+    const struct DisparityGraph& graph,
+    struct Edge edge
+);
+
+/**
+ * \brief Get neighbor pixel to current one using neighbor index.
+ *
+ * Note that the function doesn't check existence of provided neighborhood.
+ * Use ::neighborhood_exists to make sure that you use it right.
+ */
+Pixel neighbor_by_index(
+    struct Pixel pixel,
+    ULONG neighbor_index
+);
+
 #endif
