+++ lib/disparity_graph.cpp
@@ -22,6 +22,7 @@
  * SOFTWARE.
  */
 #include <disparity_graph.hpp>
+#include <indexing.hpp>
 
 #include <limits>
 #include <stdexcept>
@@ -147,185 +148,6 @@ DisparityGraph::DisparityGraph(
         this->reparametrization.end(),
         static_cast<FLOAT>(0)
     );
-}
-
-ULONG neighbor_index(
-    struct Pixel pixel,
-    struct Pixel neighbor
-)
-{
-    if (pixel.x != neighbor.x && pixel.y != neighbor.y)
-    {
-        return NEIGHBORS_COUNT;
-    }
-    if (pixel.x + 1 == neighbor.x)
-    {
-        return 0;
-    }
-    if (pixel.x == neighbor.x + 1)
-    {
-        return 1;
-    }
-    if (pixel.y + 1 == neighbor.y)
-    {
-        return 2;
-    }
-    if (pixel.y == neighbor.y + 1)
-    {
-        return 3;
-    }
-    return NEIGHBORS_COUNT;
-}
-
-bool neighborhood_exists(
-    const struct DisparityGraph& graph,
-    struct Pixel pixel,
-    struct Pixel neighbor
-)
-{
-    return neighborhood_exists_fast(
-        graph,
-        pixel,
-        neighbor_index(pixel, neighbor)
-    );
-}
-
-bool neighborhood_exists_fast(
-    const struct DisparityGraph& graph,
-    struct Pixel pixel,
-    ULONG neighbor_index
-)
-{
-    return !(neighbor_index > 3
-        || pixel.y >= graph.right.height
-        || pixel.x >= graph.right.width
-        || (pixel.x + 1 == graph.left.width && neighbor_index == 0)
-        || (pixel.x == 0 && neighbor_index == 1)
-        || (pixel.y + 1 == graph.left.height && neighbor_index == 2)
-        || (pixel.y == 0 && neighbor_index == 3)
-    );
-}
-
-bool node_exists(
-    const struct DisparityGraph& graph,
-    struct Node node
-)
-{
-    return !(
-        node.disparity >= graph.disparity_levels
-        || node.pixel.y >= graph.right.height
-        || node.pixel.x >= graph.right.width
-        || node.pixel.x + node.disparity >= graph.left.width
-    );
-}
-
-bool edge_exists(
-    const struct DisparityGraph& graph,
-    struct Edge edge
-)
-{
-    if (!node_exists(graph, edge.node))
-    {
-        return false;
-    }
-    if (!node_exists(graph, edge.neighbor))
-    {
-        return false;
-    }
-    if (!neighborhood_exists(graph, edge.node.pixel, edge.neighbor.pixel))
-    {
-        return false;
-    }
-
-    if (edge.node.pixel.y != edge.neighbor.pixel.y)
-    {
-        return true;
-    }
-
-    if (edge.node.pixel.x + 1 == edge.neighbor.pixel.x
-        && edge.node.disparity > edge.neighbor.disparity + 1)
-    {
-        return false;
-    }
-    if (edge.node.pixel.x == edge.neighbor.pixel.x + 1
-        && edge.node.disparity + 1 < edge.neighbor.disparity)
-    {
-        return false;
-    }
-
-    return true;
-}
-
-ULONG reparametrization_index_fast(
-    const struct DisparityGraph& graph,
-    struct Node node,
-    ULONG neighbor_index
-)
-{
-    ULONG index = 0;
-    index *= graph.right.width;
-    index += node.pixel.x;
-    index *= NEIGHBORS_COUNT;
-    index += neighbor_index;
-    index *= graph.disparity_levels;
-    index += node.disparity;
-    index *= graph.left.height;
-    index += node.pixel.y;
-
-    return index;
-}
-
-ULONG reparametrization_index(
-    const struct DisparityGraph& graph,
-    struct Node node,
-    struct Pixel neighbor
-)
-{
-    return reparametrization_index_fast(
-        graph,
-        node,
-        neighbor_index(node.pixel, neighbor)
-    );
-}
-
-ULONG reparametrization_index_slow(
-    const struct DisparityGraph& graph,
-    struct Edge edge
-)
-{
-    return reparametrization_index_fast(
-        graph,
-        edge.node,
-        neighbor_index(edge.node.pixel, edge.neighbor.pixel)
-    );
-}
-
-FLOAT reparametrization_value(
-    const struct DisparityGraph& graph,
-    struct Node node,
-    struct Pixel neighbor
-)
-{
-    return graph.reparametrization[reparametrization_index(graph, node, neighbor)];
-}
-
-FLOAT reparametrization_value_slow(
-    const struct DisparityGraph& graph,
-    struct Edge edge
-)
-{
-    return graph.reparametrization[reparametrization_index_slow(graph, edge)];
-}
-
-FLOAT reparametrization_value_fast(
-    const struct DisparityGraph& graph,
-    struct Node node,
-    ULONG neighbor_index
-)
-{
-    return graph.reparametrization[
-        reparametrization_index_fast(graph, node, neighbor_index)
-    ];
 }
 
 FLOAT edge_penalty(const struct DisparityGraph& graph, struct Edge edge)
